requirements.txt
shioaji==1.2.5
python-dotenv==1.0.1
boto3==1.37.3
pydantic==2.11.7

